requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0  # 兼容 Anthropic SDK
aiohttp>=3.9.0
asyncio>=3.4.3

# 视频处理
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# 工具库
tqdm>=4.66.0
loguru>=0.7.2
pydantic>=2.5.0
